Revert "kekw"
This reverts commit d93aa32a8f6f8b481e5fd19e29732caeae7d658e.

+++ 2.Eksamensprojekt/Services/AdministrationService.cs
@@ -300,8 +300,13 @@ namespace _2.Eksamensprojekt.Services
         /// Sletter en booking fra databasen baseret på id
         /// </summary>
         /// <param name="id">Typen int. Indeholder id'et på den booking der skal slettes</param>
-        public void DeleteReservation()
+        public void DeleteResevation(int id)
         {
+            if (id <= 0)
+            {
+                throw new KeyNotFoundException("Der findes ikke nogle reservationer med det ID");
+            }
+
             //string sql = "DELETE from Reservation WHERE ReservationID = @id";
             string sql = "DELETE from Reservation WHERE Dag < @nextDay";
